requirements.txt
Django==3.2.16
djangorestframework==3.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.3